+++ src/main/java/io/github/okraskat/room/optimizer/domain/OccupancyCalculatorApi.java
@@ -1,5 +1,6 @@
 package io.github.okraskat.room.optimizer.domain;
 
+import java.math.BigDecimal;
 import java.util.Collection;
 import java.util.List;
 import java.util.Map;
@@ -7,7 +8,7 @@ import java.util.Map;
 public interface OccupancyCalculatorApi {
 
     List<CalculatedOccupancy> calculate(Map<RoomCategory, Integer> availableRoomsPerCategory,
-                                        Collection<Integer> potentialPayments);
+                                        Collection<BigDecimal> potentialPayments);
 
 
 }
